Split cli.ts into per-command modules under src/commands
Each subcommand (measure, diff, detect, typecheck, review) now lives in
src/commands/<name>/index.ts and exposes a register* function that wires
itself onto the Commander program. Common parsing logic (CSV options,
exclude pattern resolution, runner auto-detection) moves to src/shared.
cli.ts becomes a thin entry point that just instantiates the program and
calls each register function.
